requirements.txt
pandas
numpy
sentence-transformers
PyPDF2
regex
openai
huggingface_hub
python-dotenv